chore (refs T22614): move DpCard to client/js.
