+++ demosplan/DemosPlanProcedureBundle/Resources/client/js/components/charts/ProcedureAnalyticsChart.vue
@@ -44,7 +44,7 @@
 </template>
 
 <script>
-import DpCard from '@DemosPlanCoreBundle/components/DpCard'
+import DpCard from '@DpJs/components/core/DpCard'
 import { DpLoading } from 'demosplan-ui/components'
 import { dpRpc } from '@DemosPlanCoreBundle/plugins/DpApi'
 import { initLineChart } from '../../lib/charts/helpers/init'
